requirements.txt
alabaster==0.7.12
astroid==2.1.0
autopep8==1.4.3
Babel==2.6.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.4.0
django-heroku==0.3.1
django-nose==1.4.6
djangorestframework==3.9.0
docutils==0.14
flake8==3.6.0
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
livereload==2.6.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mkdocs==1.0.4
nose==1.3.7
packaging==18.0
packtools==2.3.8
picles.plumber==0.11
psycopg2==2.7.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.0
pylint==2.2.2
pyparsing==2.3.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.11.0
snowballstemmer==1.2.1
sphinxcontrib-websupport==1.1.0
tornado==5.1.1
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
wrapt==1.10.11
